+++ src/main/java/team9/ddang/walk/controller/request/walk/DecisionWalkRequest.java
@@ -1,8 +1,8 @@
-package team9.ddang.walk.controller.request;
+package team9.ddang.walk.controller.request.walk;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
-import team9.ddang.walk.service.request.DecisionWalkServiceRequest;
+import team9.ddang.walk.service.request.walk.DecisionWalkServiceRequest;
 
 public record DecisionWalkRequest(
         @NotNull(message = "상대방 이메일은 필수 입력 사항입니다.")
