+++ src/main/java/team9/ddang/walk/service/WalkLocationServiceImpl.java
@@ -15,13 +15,13 @@ import team9.ddang.global.service.RedisService;
 import team9.ddang.member.entity.IsMatched;
 import team9.ddang.member.entity.Member;
 import team9.ddang.member.repository.MemberRepository;
-import team9.ddang.walk.service.request.DecisionWalkServiceRequest;
-import team9.ddang.walk.service.request.ProposalWalkServiceRequest;
-import team9.ddang.walk.service.request.StartWalkServiceRequest;
-import team9.ddang.walk.service.response.MemberNearbyInfo;
-import team9.ddang.walk.service.response.MemberNearbyResponse;
-import team9.ddang.walk.service.response.ProposalWalkResponse;
-import team9.ddang.walk.service.response.WalkWithResponse;
+import team9.ddang.walk.service.request.walk.DecisionWalkServiceRequest;
+import team9.ddang.walk.service.request.walk.ProposalWalkServiceRequest;
+import team9.ddang.walk.service.request.walk.StartWalkServiceRequest;
+import team9.ddang.walk.service.response.walk.MemberNearbyInfo;
+import team9.ddang.walk.service.response.walk.MemberNearbyResponse;
+import team9.ddang.walk.service.response.walk.ProposalWalkResponse;
+import team9.ddang.walk.service.response.walk.WalkWithResponse;
 
 import java.time.Duration;
 import java.util.ArrayList;
