+++ src/main/java/team9/ddang/walk/service/WalkService.java
@@ -1,8 +1,8 @@
 package team9.ddang.walk.service;
 
 import team9.ddang.member.entity.Member;
-import team9.ddang.walk.service.request.CompleteWalkServiceRequest;
-import team9.ddang.walk.service.response.CompleteWalkResponse;
+import team9.ddang.walk.service.request.walk.CompleteWalkServiceRequest;
+import team9.ddang.walk.service.response.walk.CompleteWalkResponse;
 
 public interface WalkService {
