+++ src/main/java/team9/ddang/walk/service/request/walk/StartWalkServiceRequest.java
@@ -1,4 +1,4 @@
-package team9.ddang.walk.service.request;
+package team9.ddang.walk.service.request.walk;
 
 import java.time.LocalDateTime;
