+++ src/main/java/team9/ddang/walk/service/response/walk/WalkWithDogInfo.java
@@ -1,4 +1,4 @@
-package team9.ddang.walk.service.response;
+package team9.ddang.walk.service.response.walk;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import team9.ddang.dog.entity.Dog;
